fixed pep issuses

+++ main.py
@@ -4,7 +4,9 @@ from server import Server
 from gameutils.game import Game
 
 if __name__ == "__main__":
-    os.remove("resources/players.txt")
+    if os.path.exists("resources/players.txt"):
+        os.remove("resources/players.txt")
+
     HOST = '127.0.0.1'
     PORT = 65432
     game = Game()
